6c935a8ad879e860273a9c01bd06f316

+++ GIT_PUSH.sh
@@ -1,7 +1,13 @@
 #!/bin/bash -e
 
 echo "----- Start -----"
-find $PWD/dist | xargs rm -rf
+
+if [ ! -d $PWD/dist ]; then
+    find $PWD/dist | xargs rm -rf
+fi
+
+npx parcel build index.html
+
 # find $PWD/* -empty | xargs rm
 git pull --ff-only
 git add .
